pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pyaudio>=0.2.14",
    "ruff>=0.15.4",
    "speechrecognition>=3.14.5",
    "ty>=0.0.19",
]
